Creating the add_item and delete_item functions:
In this step, we will create two functions, one to add items to the python to-do list and the other to delete an item from the list.
In the add_item function, the user will give it to arguments, one an Entry object and the other a Listbox object.
In this function, we will give the Entry object’s user-provided text to a variable and then insert that variable to the last of our listbox object, and the last line of our text file.
In the delete_item function, the user will give the function one argument that will be a listbox object.
In this function, we will delete the item selected in the listbox object. We will first remove that element from the listbox and then go through every text file line to see if anything matches. If it does, we remove it too.